backend/requirement.txt
fastapi 
uvicorn
numpy
xgboost
scikit-learn
tensorflow-cpu
python-multipart
pillow
python-dotenv